ics3a/encoder: move icr_encoder .exe internal API to appropriate level
